+++ src/metadata-builder/ClosureJunctionEntityMetadataBuilder.ts
@@ -42,10 +42,12 @@ export class ClosureJunctionEntityMetadataBuilder {
             entityMetadata.ownColumns.push(new ColumnMetadata({
                 connection: this.connection,
                 entityMetadata: entityMetadata,
+                closureType: "ancestor",
+                referencedColumn: primaryColumn,
                 args: {
                     target: "",
                     mode: "virtual",
-                    propertyName: "ancestor", // todo: naming strategy
+                    propertyName: primaryColumn.propertyName + "_ancestor", // todo: naming strategy
                     options: {
                         length: primaryColumn.length,
                         type: primaryColumn.type,
@@ -55,10 +57,12 @@ export class ClosureJunctionEntityMetadataBuilder {
             entityMetadata.ownColumns.push(new ColumnMetadata({
                 connection: this.connection,
                 entityMetadata: entityMetadata,
+                closureType: "descendant",
+                referencedColumn: primaryColumn,
                 args: {
                     target: "",
                     mode: "virtual",
-                    propertyName: "descendant",
+                    propertyName: primaryColumn.propertyName + "_descendant",
                     options: {
                         length: primaryColumn.length,
                         type: primaryColumn.type,