+++ test/github-issues/904/entity/Category.ts
@@ -1,12 +1,12 @@
-import {ClosureEntity} from "../../../../src/decorator/entity/ClosureEntity";
 import {PrimaryGeneratedColumn} from "../../../../src/decorator/columns/PrimaryGeneratedColumn";
 import {Column} from "../../../../src/decorator/columns/Column";
 import {TreeParent} from "../../../../src/decorator/tree/TreeParent";
 import {TreeChildren} from "../../../../src/decorator/tree/TreeChildren";
+import {Entity} from "../../../../src/decorator/entity/Entity";
+import {Tree} from "../../../../src/decorator/tree/Tree";
 
-// @Entity()
-// @Tree("closure")
-@ClosureEntity()
+@Entity("sample22_category")
+@Tree("closure-table")
 export class Category {
 
     @PrimaryGeneratedColumn()